+++ ui/components/docked-form-footer/base/example.jsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import ButtonIcon from '../../button-icons/';
-import { Popover } from '../../popovers/base/example';
+import { Popover } from '../../popovers/base-legacy/example';
 import { FeedbackHeader } from '../../popovers/error/example';
 import _ from '../../../shared/helpers';