+++ ui/components/popovers/error/example.jsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import { UtilityIcon } from '../../icons/base/example';
-import { Popover } from '../base/example';
+import { Popover } from '../base-legacy/example';
 import classNames from 'classnames';
 import _ from '../../../shared/helpers';
 
@@ -38,8 +38,8 @@ export const FeedbackHeader = props => (
 );
 
 export let FeedbackFooter = props => (
-  <div className="slds-grid slds-grid_vertical-align-center">
-    <p>Footer Item</p>
+  <div className="slds-grid slds-grid_vertical-align-center slds-grid_align-end">
+    <button className="slds-button slds-button_neutral">Learn More</button>
   </div>
 );
 
@@ -73,8 +73,35 @@ export default (
 
 export let examples = [
   {
+    id: 'with-top-nubbin',
+    label: 'Error With Top Nubbin',
+    element: (
+      <Popover
+        className="slds-popover_error slds-nubbin_top"
+        headingId={headingUniqueId}
+        header={
+          <FeedbackHeader
+            headingId={headingUniqueId}
+            title="Resolve error"
+            symbol="error"
+          />
+        }
+        closeButton
+        inverse
+      >
+        <p>
+          Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do
+          eiusmod tempor incididunt ut labore et dolore.{' '}
+          <a href="#" onClick={e => e.preventDefault()} title="Learn More">
+            Learn More
+          </a>
+        </p>
+      </Popover>
+    )
+  },
+  {
     id: 'with-footer',
-    label: 'With Footer',
+    label: 'Error With Footer',
     element: (
       <Popover
         className="slds-popover_error slds-nubbin_bottom-left"