+++ RELEASENOTES.md
@@ -40,6 +40,29 @@
 - Updated the icon in the [open as a drawer](/components/panels/#Open-as-a-drawer) example from `filterList` to `toggle_panel_left`. The new icon better represents the action evoked by the button icon.
 - Updated the icon in [drilled-in state](/components/panels/#Drilled-In-State) example that takes you back to the previous panel from `back` to `chevronleft`. This update makes the icon consistent between drilled-in states.
 
+### [Popovers](https://www.lightningdesignsystem.com/components/popovers)
+#### Changed
+- Base popovers
+    - Footer background is now explicitly set to `$color-background` design token
+    - Moved `slds-popover__header` into `slds-popover__body` and added a [Media Object](/utilities/media-objects/) wrapper to match prompt popover variant
+    - Added variant with icon
+    - Removed `slds-popover__header` bottom border
+    - Added variant with badge above header
+- Warning feedback state
+    - Footer background color is now explicitly set to transparent
+    - Body is now indented to align with header
+- Error feedback state
+    - Footer background color is now explicitly set to transparent
+    - Body is now indented to align with header
+- Prompt variant
+    - Footer top border changed to match default popover border color and width
+- Walkthrough variant
+    - Removed header background image
+    - Changed footer background color to `$palette-blue-20`
+#### Added
+- Added light and dark Brand variants, including new selectors `.slds-popover_brand, .slds-popover_brand-dark, .slds-popover_brand-top, .slds-popover_brand-left, .slds-popover_brand-bottom, .slds-popover_brand-right`
+- Added new Prompt branded variants
+
 ### [Rich Text Editor](https://www.lightningdesignsystem.com/components/rich-text-editor)
 #### Added
 - Enabled styling hooks for Rich Text Editor. See Rich Text Editor's styling hooks overview table for a full listing of the currently available hooks.